BUG: rename symbols for iOS compatibility.
Apple forbids some symbols that Numpy is linked against (ccopy, dcopy, scopy, zcopy and xerbla).
Compilation and tests work fine but the App Store complains and prevents uploading such builds.

This patch aliases the culprits to the public, allowed method names, allowing the linking to
proceed as before and the resulting binary to be accepted by the App Store.